+++ Module.php
@@ -6,7 +6,7 @@ namespace wdmg\api;
  * Yii2 API
  *
  * @category        Module
- * @version         1.3.7
+ * @version         1.3.8
  * @author          Alexsander Vyshnyvetskyy <alex.vyshnyvetskyy@gmail.com>
  * @link            https://github.com/wdmg/yii2-api
  * @copyright       Copyright (c) 2019 - 2020 W.D.M.Group, Ukraine
@@ -45,7 +45,7 @@ class Module extends BaseModule
     /**
      * @var string the module version
      */
-    private $version = "1.3.7";
+    private $version = "1.3.8";
 
     /**
      * @var integer, priority of initialization
@@ -93,6 +93,7 @@ class Module extends BaseModule
             "wdmg\api\models\api\NewsAPI" => true,
             "wdmg\api\models\api\BlogAPI" => true,
             "wdmg\api\models\api\OptionsAPI" => false,
+            "wdmg\api\models\api\ContentAPI" => true,
             "wdmg\api\models\api\PagesAPI" => true,
             "wdmg\api\models\api\MediaAPI" => true,
             "wdmg\api\models\api\SearchAPI" => true,
@@ -111,6 +112,7 @@ class Module extends BaseModule
             "wdmg\api\models\api\NewsAPI" => true,
             "wdmg\api\models\api\BlogAPI" => true,
             "wdmg\api\models\api\OptionsAPI" => true,
+            "wdmg\api\models\api\ContentAPI" => true,
             "wdmg\api\models\api\PagesAPI" => true,
             "wdmg\api\models\api\MediaAPI" => true,
             "wdmg\api\models\api\SearchAPI" => true,